backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-multipart
pydantic
pydantic-settings
dashscope
chromadb
Pillow
numpy
opencv-python-headless
ultralytics
sahi
httpx
aiofiles
python-dotenv
websockets
